osd: skip dup setallochint op
when the object is exists, the setallochint op should have been set
by previous op, so skip this.

We should make sure that when the object is not exists, setallochint
has been called before.

this small change would reduce latency in do_transaction by 8-10 us
on pcie ssd.

ANGEL-34 #finish
Signed-off-by: Xinze Chi <xinze@xsky.com>